refactor: patch alarm token logic AlarmRepository로 이동
