+++ data/src/main/java/com/asap/data/remote/datasource/AlarmRemoteDataSource.kt
@@ -1,5 +1,6 @@
 package com.asap.data.remote.datasource
 
+import com.asap.data.remote.request.PatchAlarmTokenBody
 import com.asap.data.remote.service.AlarmService
 import com.asap.domain.entity.remote.AlarmOffRequestBody
 import com.asap.domain.entity.remote.WhetherResponse
@@ -12,6 +13,15 @@ import javax.inject.Inject
 class AlarmRemoteDataSource @Inject constructor(
     private val alarmService: AlarmService
 ) {
+    suspend fun patchAlarmToken(userId: Int, token: String): WhetherResponse?  {
+        return PatchAlarmTokenBody(userId = userId, alarmToken = token).let { body ->
+            alarmService.patchAlarmToken(body)
+        }.run {
+            if (!isSuccessful) throw HttpException(this)
+            body().also(::println)
+        }
+    }
+
     suspend fun fetchAlarmOffRate(userId: String): Flow<AlarmOffRate?> = flow {
         val response = alarmService.fetchAlarmOffRate(userId = userId)
         if (!response.isSuccessful) {
