+++ data/src/main/java/com/asap/data/remote/service/AuthService.kt
@@ -9,13 +9,9 @@ import com.asap.domain.entity.remote.auth.RefreshTokenResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.Header
-import retrofit2.http.PATCH
 import retrofit2.http.POST
 
 interface AuthService {
-    @PATCH("/admin/alarm_tokne")
-    suspend fun patchAlarmToken(@Body body: PatchAlarmTokenBody): Response<WhetherResponse>
-
     @POST("/auth/kakao")
     suspend fun authKakao(@Body body: AuthKakaoBody): Response<AuthKakaoResponse?>
