+++ data/src/main/java/com/asap/data/repository/AlarmRepositoryImpl.kt
@@ -6,8 +6,10 @@ import com.asap.domain.entity.local.AlarmEntity
 import com.asap.domain.entity.remote.AlarmSummary
 import com.asap.domain.entity.remote.WhetherResponse
 import com.asap.domain.entity.remote.alarm.AlarmOffRate
+import com.asap.domain.entity.remote.auth.TokenManager
 import com.asap.domain.repository.AlarmRepository
 import kotlinx.coroutines.flow.Flow
+import retrofit2.HttpException
 import javax.inject.Inject
 
 class AlarmRepositoryImpl @Inject constructor(
@@ -16,6 +18,16 @@ class AlarmRepositoryImpl @Inject constructor(
 ) : AlarmRepository{
     private val userDao = localDataSource.userDao()
     private val dao = localDataSource.deactivatedAlarmDao()
+
+    override suspend fun patchAlarmToken(token: String): Boolean {
+        return try {
+            val userId = (userDao.selectAll().firstOrNull()?.userId ?: "-1").toInt()
+            remoteDataSource.patchAlarmToken(userId, TokenManager.fcmToken)?.result ?: false
+        } catch (e: HttpException) {
+            println(e)
+            false
+        }
+    }
 
     override suspend fun fetchAlarmOffRate(): Flow<AlarmOffRate?> {
         val userId = userDao.selectAll().firstOrNull()?.userId ?: "-1"
