+++ domain/src/main/java/com/asap/domain/di/UseCaseModule.kt
@@ -10,6 +10,8 @@ import com.asap.domain.usecase.alarm.FetchAlarmOffRateUseCase
 import com.asap.domain.usecase.alarm.FetchAlarmOffRateUseCaseImpl
 import com.asap.domain.usecase.alarm.GetDeactivatedAlarmListUseCase
 import com.asap.domain.usecase.alarm.GetDeactivatedAlarmlistUseCaseImpl
+import com.asap.domain.usecase.alarm.PatchAlarmTokenUseCase
+import com.asap.domain.usecase.alarm.PatchAlarmTokenUseCaseImpl
 import com.asap.domain.usecase.auth.AuthKakaoUseCase
 import com.asap.domain.usecase.auth.AuthKakaoUseCaseImpl
 import com.asap.domain.usecase.auth.CacheAuthKakaoUseCaseImpl
@@ -18,8 +20,6 @@ import com.asap.domain.usecase.auth.CheckAuthUseCaseImpl
 import com.asap.domain.usecase.auth.CheckCachedAuthUseCase
 import com.asap.domain.usecase.auth.KakaoLoginUseCase
 import com.asap.domain.usecase.auth.KakaoLoginUseCaseImpl
-import com.asap.domain.usecase.auth.PatchAlarmTokenUseCase
-import com.asap.domain.usecase.auth.PatchAlarmTokenUseCaseImpl
 import com.asap.domain.usecase.auth.RefreshTokenUseCase
 import com.asap.domain.usecase.auth.RefreshTokenUseCaseImpl
 import com.asap.domain.usecase.auth.RegisterTokenUseCase
@@ -88,9 +88,6 @@ interface UseCaseModule {
 
     @Binds
     fun bindCheckCachedAuthUseCase(usecaseImpl: CheckAuthUseCaseImpl): CheckCachedAuthUseCase
-
-    @Binds
-    fun bindPatchAlarmTokenUseCase(usecaseImpl: PatchAlarmTokenUseCaseImpl): PatchAlarmTokenUseCase
 
     /**
      * User usecase
@@ -191,6 +188,9 @@ interface UseCaseModule {
     /**
      * Alarm usecases
      */
+    @Binds
+    fun bindPatchAlarmTokenUseCase(usecaseImpl: PatchAlarmTokenUseCaseImpl): PatchAlarmTokenUseCase
+
     @Binds
     fun bindGetDeactivatedAlarmListUseCase(
         getDeactivatedAlarmlistUseCaseImpl: GetDeactivatedAlarmlistUseCaseImpl
