+++ domain/src/main/java/com/asap/domain/repository/AlarmRepository.kt
@@ -7,6 +7,9 @@ import com.asap.domain.entity.remote.alarm.AlarmOffRate
 import kotlinx.coroutines.flow.Flow
 
 interface AlarmRepository {
+    // FCM token 동기화
+    suspend fun patchAlarmToken(token: String): Boolean
+
     // 알람 해제율 조회
     suspend fun fetchAlarmOffRate(): Flow<AlarmOffRate?>
