+++ domain/src/main/java/com/asap/domain/repository/AuthRepository.kt
@@ -21,7 +21,5 @@ interface AuthRepository {
 
     suspend fun refreshToken(): Boolean
 
-    suspend fun patchAlarmToken(token: String): Boolean
-
     suspend fun updateToken(accessToken: String, refreshToken: String)
 }
