+++ presentation/.idea/.gitignore
@@ -1,0 +1,3 @@
+# 디폴트 무시된 파일
+/shelf/
+/workspace.xml
